+++ packages/plugin-abi/src/services/contract.ts
@@ -1,7 +1,7 @@
 import { elizaLogger } from "@elizaos/core";
 import type { Account, Address, PublicClient, WalletClient } from "viem";
 import { WalletService } from "./wallet.js";
-import { withRetry } from "../lib/helpers.js";
+import { withRetry, formatResult } from "../lib/helpers.js";
 import type { Chain } from "viem/chains";
 
 export class ContractService {
@@ -42,34 +42,10 @@ export class ContractService {
 					functionName,
 					args,
 				});
-				return this.formatResult(res);
+				return formatResult(res);
 			},
 			{ logPrefix: `Read ${functionName}` },
 		);
-	}
-
-	private formatResult(result: any) {
-		// If it's a BigInt, convert to string
-		if (typeof result === "bigint") {
-			return result.toString();
-		}
-
-		// If it's an array, format each element
-		if (Array.isArray(result)) {
-			return result.map((item) => this.formatResult(item));
-		}
-
-		// If it's an object, format each property
-		if (result && typeof result === "object") {
-			const formatted = {};
-			for (const key in result) {
-				formatted[key] = this.formatResult(result[key]);
-			}
-			return formatted;
-		}
-
-		// Otherwise return as is
-		return result;
 	}
 
 	async callWriteFunction(functionName: string, args: any[] = []) {
